#52 Add support for the timestamp custom function
Also it is the time to introduce the testing framework. Tried Spek, but it is the hustle to setup it (i.e. lots of magic), and it is far from power and elegance of Spock.

Although official kotlin-test looks much nicer and closer to Spock. So will have to switch to kotlin-test instead.