#135 Enhance the logic of extracting the data equal to the max page size
As per documentation the filtering occurs after query/scan, so this explains why, for example, only 7 records are returned instead of 100.

So need to enhance the QueryResult to fetch the data up to the maxPageSize or less if no more data is available. Also require to update the page num calculation function.