#77 Use Caffeine cache to keep 10 most often used predicates in the memory
Although not sure whether it gives memory or performance benefit, but at least I don't like the idea to create new predicate on every character typing and better to reuse some of them.

Also don't know whether the size of 10 is the good default size or need to use bigger one.