#4 Enable kotlin serialization runtime to work
Although it already looks quite limiting to need to open the properties, and define them and override in the child classes.

Maybe there are better ways to handle this.

Another topic will arise on the discriminator property, whether the framework supports this, otherwise might switch to the 3rd party library.