#192 Add refresh parameter to the list names method to avoid fetching the list from the dist over and over again
It is lesser the issue for the session names, as they are actually fetched only once, but it is kind of the issue for the custom queries list, where we have to fetch the list for every query/scan data result. So the next step is actually to cache the data and only refresh it from the dist when the refresh is set.